+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [Convert-CharCase][] the corresponding delimiter with the -AsSentence switch.
 - Added [Convert-CharCase][] parameters -ExcludePattern, -MatchPattern.
 - Added [Replace-InQuote][] -FirstMatch option.
+- Added [Script-InQuote][] function.
 
 ## [0.8.0] - 2025-04-27
 
@@ -314,6 +315,8 @@ Breaking Changes
 [UnZip-GzFile]: src/UnZip-GzFile_function.ps1
 
 [Convert-CharCase]: src/Convert-CharCase_function.ps1
+
+[Script-InQuote]: src/Script-InQuote_function.ps1
 
 [unreleased]: https://github.com/btklab/pwsh-sketches/compare/0.8.0..HEAD
 [0.8.0]: https://github.com/btklab/pwsh-sketches/releases/tag/0.8.0